Исправил страницы входа и регистрации
