+++ src/app/App.tsx
@@ -1,26 +1,24 @@
-import { Suspense } from 'react';
 import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom';
-import { AuthContainer } from '@processes';
-import { MainPage, ProfilePage, RegistrationPage, LoginPage } from '@pages';
+import { Container } from '@processes';
+// import { MainPage, ProfilePage, RegistrationPage, LoginPage } from '@pages';
+import { RegistrationPage, LoginPage } from '@pages';
 import { routes, AppProvider } from '@shared';
 import './app.css';
 
 export const App = () => {
   return (
     <AppProvider>
-      <Suspense fallback={<></>}>
-        <BrowserRouter>
-          <Routes>
-            <Route path="auth" element={<AuthContainer />}>
-              <Route index element={<LoginPage />} />
-              <Route path="registration" element={<RegistrationPage />} />
-            </Route>
-            <Route path={routes.main} element={<MainPage />} />
-            <Route path={routes.profile} element={<ProfilePage />} />
-            <Route path="*" element={<Navigate to={routes.login} replace />} />
-          </Routes>
-        </BrowserRouter>
-      </Suspense>
+      <BrowserRouter>
+        <Routes>
+          <Route path="" element={<Container />}>
+            <Route path={routes.login} element={<LoginPage />} />
+            <Route path={routes.registration} element={<RegistrationPage />} />
+            {/* <Route path={routes.main} element={<MainPage />} />
+              <Route path={routes.profile} element={<ProfilePage />} /> */}
+          </Route>
+          <Route path="*" element={<Navigate to={routes.main} replace />} />
+        </Routes>
+      </BrowserRouter>
     </AppProvider>
   );
 };
