+++ src/pages/login-page/login-page.module.scss
@@ -1,9 +1,3 @@
-.background {
-  min-height: 100vh;
-  width: 100%;
-  background-color: var(--grey);
-}
-
 .window {
   width: 360px;
   padding: 25px;
@@ -13,7 +7,7 @@
   background-color: var(--main);
 }
 
-@media screen and (max-width: 1024px) {
+@media screen and (max-width: 768px) {
   .window {
     width: 90%;
     max-width: 350px;
