+++ src/pages/registration-page/index.ts
@@ -1,0 +1,1 @@
+export { RegistrationPage } from './registration-page';
