+++ src/pages/registration-page/registration-page.module.scss
@@ -1,0 +1,18 @@
+.window {
+  width: 360px;
+  padding: 25px;
+  border-radius: 17px;
+  margin: 0 auto;
+  transform: translateY(40px);
+  background-color: var(--main);
+}
+
+@media screen and (max-width: 768px) {
+  .window {
+    width: 90%;
+    max-width: 350px;
+    padding: 12px;
+    border-radius: 3px;
+    transform: translateY(30px);
+  }
+}
