+++ src/pages/registration-page/registration-page.tsx
@@ -1,45 +1,15 @@
-import { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { useEffect } from 'react';
 import { RegisterForm } from '@widgets';
-import { routes, url, Loading } from '@shared';
+import styles from './registration-page.module.scss';
 
-function RegistrationPage() {
-  const navigate = useNavigate(),
-    [loading, setLoading] = useState(true);
-
+export const RegistrationPage = () => {
   useEffect(() => {
     document.title = 'Регистрация';
-
-    async function checkLogin() {
-      const response = await fetch(`${url}api/user/check_login`, {
-        method: 'GET',
-        credentials: 'include',
-      });
-
-      if (response.ok) {
-        setLoading(false);
-        navigate(routes.main);
-      } else {
-        setLoading(false);
-      }
-    }
-
-    checkLogin();
   }, []);
 
-  const render = () => {
-    if (loading) {
-      return (
-        <>
-          <Loading size="max" />
-        </>
-      );
-    } else {
-      return <RegisterForm />;
-    }
-  };
-
-  return <>{render()}</>;
-}
-
-export default RegistrationPage;
+  return (
+    <div className={styles.window}>
+      <RegisterForm />
+    </div>
+  );
+};
