--- src/processes/auth-container/index.ts
@@ -1,1 +1,0 @@
-export { AuthContainer } from './ui/auth-container';
