+++ src/processes/container/container.module.scss
@@ -1,0 +1,5 @@
+.background {
+  min-height: 100vh;
+  width: 100%;
+  background-color: var(--grey);
+}
