+++ src/shared/config/config.ts
@@ -1,8 +1,8 @@
 export const routes = {
-  login: '/auth',
-  registration: '/auth/registration',
+  login: '/login',
+  registration: '/registration',
   main: '/',
   profile: '/profile',
 };
 
-export const url = 'http://127.0.0.1:5000/';
+export const URL = 'http://127.0.0.1:5000/';
