+++ src/shared/ui/confirm-button/confirm-button.module.scss
@@ -1,0 +1,37 @@
+.button {
+  width: 100%;
+  padding: 0 10px;
+  font-size: 18px;
+  height: 50px;
+  border-radius: 25px;
+  color: #fff;
+  background: var(--grad);
+  border: none;
+  user-select: none;
+
+  &:hover {
+    cursor: pointer;
+  }
+}
+
+.loading {
+  opacity: 0.6;
+
+  &:hover {
+    cursor: default;
+  }
+}
+
+.confuse {
+  width: 100%;
+  font-size: 15px;
+  height: 50px;
+  color: #fff;
+  border-radius: 5px;
+  background: none;
+  user-select: none;
+  border: 1px solid #fff;
+  display: flex;
+  align-items: center;
+  text-align: center;
+}
