--- src/shared/ui/confirm-button/ui/confirm-button.module.scss
@@ -1,19 +1,0 @@
-.button {
-  width: 100%;
-  padding: 0 10px;
-  font-size: 18px;
-  height: 50px;
-  border-radius: 25px;
-  color: #fff;
-  background: var(--grad);
-  border: none;
-  user-select: none;
-
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-.loading {
-  opacity: 0.6;
-}
