+++ src/shared/ui/data-input/data-input.module.scss
@@ -32,8 +32,7 @@
   color: #f00;
 }
 
-
-@media screen and (max-width: 1024px) {
+@media screen and (max-width: 768px) {
   .input {
     font-size: 16px;
   }
