+++ src/widgets/login-form/login-form.module.scss
@@ -20,7 +20,7 @@
   font-size: 16px;
 }
 
-@media screen and (max-width: 1024px) {
+@media screen and (max-width: 768px) {
   .headline {
     font-size: 24px;
   }
