+++ src/widgets/registration-form/registration-form.module.scss
@@ -1,0 +1,35 @@
+.headline {
+  font-size: 30px;
+  font-weight: bold;
+  text-align: center;
+  margin-bottom: 30px;
+  color: #fff;
+}
+
+.inputs {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.link {
+  display: block;
+  color: #fff;
+  text-decoration: none;
+  margin: 25px 0 15px 0;
+  font-size: 16px;
+}
+
+@media screen and (max-width: 768px) {
+  .headline {
+    font-size: 24px;
+  }
+
+  .inputs {
+    gap: 10px;
+  }
+
+  .link {
+    font-size: 15px;
+  }
+}
