--- src/widgets/registration-form/ui/registration-form.module.scss
@@ -1,30 +1,0 @@
-.container {
-  width: 90%;
-  max-width: 400px;
-  padding: 10px;
-  border: 3px solid var(--main);
-  border-radius: 8px;
-  margin: 20px auto;
-}
-
-.headline {
-  font-size: 24px;
-  font-weight: bold;
-  text-align: center;
-  margin-bottom: 30px;
-}
-
-.inputs {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-  margin-bottom: 10px;
-}
-
-.link {
-  display: block;
-  color: var(--main);
-  text-decoration: none;
-  margin-bottom: 10px;
-  font-size: 14px;
-}
